+++ src/engine/engine_collision_gjk.h
@@ -20,7 +20,7 @@
 
 #include <mujoco/mjexport.h>
 #include <mujoco/mjmodel.h>
-#include <mujoco/mjtnum.h>
+#include <mujoco/mjtype.h>
 
 #include "engine/engine_collision_convex.h"